Added getHandValue to Player.py
added the function to calculate the player's hand, will return the highest number below 21 that the players cards can make (aces are 1 or 11). Will return the string Bust when a player can only make values above 21 with their cards and the string Blackjack when the two starting cards they get equal 21 (a natural blackjack).